feat: store money as minor units and make checkout idempotent

+++ internal/models/order_item.go
@@ -9,7 +9,7 @@ type OrderItem struct {
 	// Product     Product   `gorm:"foreignKey:ProductID" json:"product"`
 	ProductName string    `gorm:"not null" json:"product_name"`
 	Quantity    int       `gorm:"not null" json:"quantity"`
-	Price       float64   `gorm:"not null" json:"price"`
+	Price       int64     `gorm:"not null" json:"price"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
